Rename customHooks folder to simply hooks folder
