+++ src/components/DAO/CreateDaoAdminProposal/CreateAuction.tsx
@@ -2,7 +2,7 @@ import {ChangeEvent, FunctionComponent, useContext, useState} from "react"
 import Input from "../../Controls/Input"
 import Select from "../../Controls/Select"
 import Button from "../../Controls/Button"
-import useNFTs from "../../../customHooks/getters/useNFTs"
+import useNFTs from "../../../hooks/getters/useNFTs"
 import ErrorPlaceholder from "../../ErrorPlaceholder"
 import Loader from "../../Loader"
 import {NFT} from "../../../types/NFT"
