+++ src/components/DAO/CreateDaoAdminProposal/index.tsx
@@ -6,7 +6,7 @@ import {AuthContext} from "../../../context/AuthContext"
 import {SafeProposalsTypeNames, SafeProposalType} from "../../../types/safeProposal"
 import CreateAuction from "./CreateAuction"
 import CancelAuction from "./CancelAuction"
-import useDAOProposals from "../../../customHooks/getters/useDAOProposals"
+import useDAOProposals from "../../../hooks/getters/useDAOProposals"
 import ErrorPlaceholder from "../../ErrorPlaceholder"
 import Loader from "../../Loader"
 import GeneralEVM from "./GeneralEVM"
