+++ src/components/Modals/CreateNFTModal/index.tsx
@@ -17,7 +17,7 @@ import "./styles.scss"
 import {toastError} from "../../Toast"
 import addDAONFT from "../../../api/firebase/NFT/addDAONFT"
 import transferNFT from "../../../api/ethers/functions/NFT/transferNFT"
-import useUser from "../../../customHooks/getters/useUser"
+import useUser from "../../../hooks/getters/useUser"
 import {Domain} from "../../../types/user"
 import Loader from "../../Loader"
 import ErrorPlaceholder from "../../ErrorPlaceholder"
