+++ src/components/Modals/DecentralizeDAOModal/index.tsx
@@ -4,7 +4,7 @@ import Modal from "../Modal"
 import RadioButton from "../../Controls/RadioButton"
 import CreateERC20Token from "../../DAO/CreateERC20Token"
 import DecentralizeDAO from "../../DAO/DecentralizeDAO"
-import useDAO from "../../../customHooks/getters/useDAO"
+import useDAO from "../../../hooks/getters/useDAO"
 import Loader from "../../Loader"
 import ErrorPlaceholder from "../../ErrorPlaceholder"
 import "./styles.scss"
