+++ src/components/NFTGallery/index.tsx
@@ -4,7 +4,7 @@ import Select from "../Controls/Select"
 import Gallery from "../Gallery"
 import Button from "../Controls/Button"
 import {NFTSnapshot, NftSort} from "../../types/NFT"
-import useNFTs from "../../customHooks/getters/useNFTs"
+import useNFTs from "../../hooks/getters/useNFTs"
 import ErrorPlaceholder from "../ErrorPlaceholder"
 import Loader from "../Loader"
 import SearchIcon from "../../icons/SearchIcon"
