+++ src/components/User/ProfileDAOs/index.tsx
@@ -2,7 +2,7 @@ import {FunctionComponent, useContext} from "react"
 import Input from "../../Controls/Input"
 import Table from "../../Table"
 import "./styles.scss"
-import useMyDAOs from "../../../customHooks/getters/useMyDAOs"
+import useMyDAOs from "../../../hooks/getters/useMyDAOs"
 import Loader from "../../Loader"
 import ErrorPlaceholder from "../../ErrorPlaceholder"
 import {AuthContext} from "../../../context/AuthContext"
