+++ src/pages/DAO/index.tsx
@@ -2,7 +2,7 @@ import {FunctionComponent, useContext, useState} from "react"
 import {useParams} from "react-router-dom"
 import HorizontalMenu from "../../components/HorizontalMenu"
 import "./styles.scss"
-import useDAO from "../../customHooks/getters/useDAO"
+import useDAO from "../../hooks/getters/useDAO"
 import ErrorPlaceholder from "../../components/ErrorPlaceholder"
 import Loader from "../../components/Loader"
 import {AuthContext} from "../../context/AuthContext"
