+++ src/pages/Profile/index.tsx
@@ -8,7 +8,7 @@ import ProfileEdit from "../../components/User/ProfileEdit"
 import ProfileDAOs from "../../components/User/ProfileDAOs"
 import CreateCustomDomainModal from "../../components/Modals/CreateCustomDomainModal"
 import CreateNFTModal from "../../components/Modals/CreateNFTModal"
-import useUser from "../../customHooks/getters/useUser"
+import useUser from "../../hooks/getters/useUser"
 import ErrorPlaceholder from "../../components/ErrorPlaceholder"
 import Loader from "../../components/Loader"
 import updateUserImage from "../../api/firebase/user/updateUserImage"
